requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aioredis==2.0.1
aiorq==1.1.9
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
black==24.4.2
certifi==2024.7.4
click==8.1.7
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
mypy-extensions==1.0.0
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pandas-stubs==2.2.2.240603
pathspec==0.12.1
platformdirs==4.2.2
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.7
rq==1.16.2
six==1.16.0
types-pytz==2024.1.0.20240417
typing_extensions==4.12.2
tzdata==2024.1
yarl==1.9.4
